@@ -1,38 +1,33 @@
 {
   "name": "ypf-methods",
-  "description": "全局混入vue的js方法包",
-  "version": "1.2.5",
+  "description": "常用基础工具方法",
+  "version": "1.3.21",
   "author": "ypf",
   "main": "dist/build.js",
   "license": "MIT",
   "scripts": {
-    "dev": "cross-env NODE_ENV=development webpack-dev-server --open --hot",
-    "build": "cross-env NODE_ENV=production webpack --progress --hide-modules"
+    "build": "webpack --config webpack.config.js",
+    "goNpm": "npm run build && npm login && npm publish"
   },
   "keywords": [
     "ypf",
     "ypf-methods"
   ],
-  "dependencies": {
-    "vue": "^2.5.11"
-  },
   "browserslist": [
     "> 1%",
     "last 2 versions",
     "not ie <= 8"
   ],
+  "dependencies": {},
   "devDependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-preset-env": "^1.6.0",
+    "@babel/plugin-transform-runtime": "^7.12.1",
+    "@babel/polyfill": "^7.12.1",
+    "babel-core": "^6.26.3",
+    "babel-loader": "^7.1.5",
+    "babel-preset-env": "^1.7.0",
     "babel-preset-stage-3": "^6.24.1",
-    "cross-env": "^5.0.5",
-    "css-loader": "^0.28.7",
-    "file-loader": "^1.1.4",
-    "html-webpack-plugin": "^3.2.0",
-    "vue-loader": "^13.0.5",
-    "vue-template-compiler": "^2.4.4",
-    "webpack": "^3.6.0",
-    "webpack-dev-server": "^2.9.1"
+    "path": "^0.12.7",
+    "webpack": "^5.7.0",
+    "webpack-cli": "^4.2.0"
   }
 }
